+++ pkg/api/login.go
@@ -13,41 +13,6 @@ type LoginHandler struct {
 	NavBar
 }
 
-const SessionCookie = "vvgo_session"
-
-var sessions = new(Sessions)
-
-type Sessions struct {
-	sessions []Session
-}
-
-func (x *Sessions) Add(session Session) {
-	x.sessions = append(x.sessions, session)
-}
-
-func (x *Sessions) Read(r *http.Request) (Session, bool) {
-	cookie, err := r.Cookie(SessionCookie)
-	if err != nil {
-		logger.WithError(err).Debug("cookie error")
-		return Session{}, false
-	}
-	return sessions.Get(cookie.Value)
-}
-
-func (x *Sessions) Get(key string) (Session, bool) {
-	for _, session := range x.sessions {
-		if session.Key == key {
-			return session, true
-		}
-	}
-	return Session{}, false
-}
-
-type Session struct {
-	Key       string
-	VVVGOUser string
-}
-
 func (x LoginHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	session, ok := sessions.Read(r)
 	if !ok {
