--- pkg/storage/sessions.go
@@ -1,3 +1,0 @@
-package storage
-
-
